Documentation update. Bump version to 0.1.0-dev
