+++ doc/features.md
@@ -1,0 +1,17 @@
+# Features
+
+* [Algebra](./algebra.md) - [Context-based](./contexts.md) operations on different primitives and structures.
+
+* [NDStructures](./nd-structure.md)
+
+* [Linear algebra](linear) - Matrices, operations and linear equations solving. To be moved to separate module. Currently supports basic
+api and multiple library back-ends.
+
+* [Histograms](./histograms.md) - Multidimensional histogram calculation and operations.
+
+* [Expressions](./expressions.md)
+
+* Commons math integration
+
+* Koma integration
+
