+++ doc/nd-structure.md
@@ -1,3 +1,7 @@
+# Nd-structure generation and operations
+
+**TODO**
+
 # Performance for n-dimensional structures operations
 
 One of the most sought after features of mathematical libraries is the high-performance operations on n-dimensional
