+++ kmath-memory/src/commonMain/kotlin/scientifik/memory/MemorySpec.kt
@@ -1,5 +1,7 @@
 package scientifik.memory
 
+import kotlin.reflect.KClass
+
 /**
  * A specification to read or write custom objects with fixed size in bytes
  */
